pyproject.toml
[tool.poetry]
name = "backlog-app"
version = "0.1.0"
description = ""
authors = ["Manuel Luis González <manuel.gonzalez@itcl.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
numpy = "^1.26.4"
streamlit = "^1.31.1"
plotly = "^5.19.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
